fix: resolve quick-fix review issues
Extract closure-capture helpers, split misplaced map docs, share MCP soak test helpers, and use an ExUnit skip for the stdio release soak.

Verified with focused root Lisp tests, catalog builtins tests, MCP soak tests with PTC_SOAK_ITERATIONS=3, format checks, and git diff --check.
