+++ lib/ptc_runner/lisp/eval.ex
@@ -17,6 +17,7 @@ defmodule PtcRunner.Lisp.Eval do
 
   require Logger
 
+  alias PtcRunner.Lisp.ClosureCapture
   alias PtcRunner.Lisp.CoreAST
   alias PtcRunner.Lisp.Env
   alias PtcRunner.Lisp.Eval.{Apply, Patterns}
@@ -1449,7 +1450,7 @@ defmodule PtcRunner.Lisp.Eval do
     # session's TTL. The collector is scope-aware so params, named-fn self
     # bindings, and inner let/fn/loop bindings don't cause an unrelated outer
     # value with the same name to be captured.
-    referenced = referenced_vars(body, params, extra_bound_names)
+    referenced = ClosureCapture.referenced_vars(body, params, extra_bound_names)
 
     # Keep an entry only if the body references it AND it is EITHER:
     #   * a key in `locals` (let/fn-param, possibly shadowing a builtin like
@@ -1498,125 +1499,6 @@ defmodule PtcRunner.Lisp.Eval do
       {:ok, atom} -> [name, atom]
       :error -> [name]
     end
-  end
-
-  # Collects the free `{:var, _}` names in a CoreAST subtree, normalized to
-  # strings. Known binding forms are handled with their lexical scope; any
-  # unrecognized tuple/list/map is still recursed into so a future AST node
-  # cannot silently cause a referenced var to be missed.
-  @spec referenced_vars(term(), CoreAST.fn_params(), [CoreAST.name()]) :: MapSet.t(String.t())
-  defp referenced_vars(ast, params, extra_bound_names) do
-    initial_bound =
-      params
-      |> bound_names_from_params()
-      |> MapSet.union(normalize_name_set(extra_bound_names))
-
-    collect_free_var_refs(ast, MapSet.new(), initial_bound)
-  end
-
-  defp collect_free_var_refs({:var, name}, acc, bound) do
-    name = to_string(name)
-
-    if MapSet.member?(bound, name) do
-      acc
-    else
-      MapSet.put(acc, name)
-    end
-  end
-
-  defp collect_free_var_refs({:let, bindings, body}, acc, bound) do
-    collect_free_refs_in_bindings(bindings, body, acc, bound)
-  end
-
-  defp collect_free_var_refs({:loop, bindings, body}, acc, bound) do
-    collect_free_refs_in_bindings(bindings, body, acc, bound)
-  end
-
-  defp collect_free_var_refs({:fn, params, body}, acc, bound) do
-    collect_free_var_refs(body, acc, MapSet.union(bound, bound_names_from_params(params)))
-  end
-
-  defp collect_free_var_refs({:fn, name, params, body}, acc, bound) do
-    inner_bound =
-      bound
-      |> MapSet.union(bound_names_from_params(params))
-      |> MapSet.put(to_string(name))
-
-    collect_free_var_refs(body, acc, inner_bound)
-  end
-
-  defp collect_free_var_refs(tuple, acc, bound) when is_tuple(tuple) do
-    collect_free_var_refs(Tuple.to_list(tuple), acc, bound)
-  end
-
-  defp collect_free_var_refs(list, acc, bound) when is_list(list) do
-    Enum.reduce(list, acc, &collect_free_var_refs(&1, &2, bound))
-  end
-
-  defp collect_free_var_refs(map, acc, bound) when is_map(map) and not is_struct(map) do
-    Enum.reduce(map, acc, fn {k, v}, inner ->
-      v
-      |> collect_free_var_refs(collect_free_var_refs(k, inner, bound), bound)
-    end)
-  end
-
-  defp collect_free_var_refs(_other, acc, _bound), do: acc
-
-  defp collect_free_refs_in_bindings(bindings, body, acc, bound) do
-    {acc, bound} =
-      Enum.reduce(bindings, {acc, bound}, fn {:binding, pattern, value_ast},
-                                             {inner_acc, inner_bound} ->
-        inner_acc = collect_free_var_refs(value_ast, inner_acc, inner_bound)
-        inner_bound = MapSet.union(inner_bound, bound_names_from_pattern(pattern))
-        {inner_acc, inner_bound}
-      end)
-
-    collect_free_var_refs(body, acc, bound)
-  end
-
-  defp bound_names_from_params(params) when is_list(params) do
-    params
-    |> Enum.flat_map(&names_from_pattern/1)
-    |> normalize_name_set()
-  end
-
-  defp bound_names_from_params({:variadic, leading, rest_pattern}) do
-    (Enum.flat_map(leading, &names_from_pattern/1) ++ names_from_pattern(rest_pattern))
-    |> normalize_name_set()
-  end
-
-  defp bound_names_from_pattern(pattern) do
-    pattern
-    |> names_from_pattern()
-    |> normalize_name_set()
-  end
-
-  defp names_from_pattern({:var, name}), do: [name]
-  defp names_from_pattern({:destructure, {:keys, keys, _defaults}}), do: keys
-
-  defp names_from_pattern({:destructure, {:map, keys, renames, _defaults}}) do
-    keys ++
-      Enum.flat_map(renames, fn {target_pattern, _source_key} ->
-        names_from_pattern(target_pattern)
-      end)
-  end
-
-  defp names_from_pattern({:destructure, {:as, name, inner}}),
-    do: [name | names_from_pattern(inner)]
-
-  defp names_from_pattern({:destructure, {:seq, patterns}}),
-    do: Enum.flat_map(patterns, &names_from_pattern/1)
-
-  defp names_from_pattern({:destructure, {:seq_rest, leading, rest}}) do
-    Enum.flat_map(leading, &names_from_pattern/1) ++ names_from_pattern(rest)
-  end
-
-  defp names_from_pattern(_other), do: []
-
-  defp normalize_name_set(names) do
-    names
-    |> Enum.map(&to_string/1)
-    |> MapSet.new()
   end
 
   # Only embed captured_locals in meta when non-empty — keeps closure size
