+++ mcp_server/test/soak/many_turns_soak_test.exs
@@ -33,31 +33,20 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
   """
   use ExUnit.Case, async: false
 
-  alias PtcRunnerMcp.{ConcurrencyGate, Sessions, Tools}
-  alias PtcRunnerMcp.Sessions.Config, as: SessionsConfig
   alias PtcRunnerMcp.Sessions.Registry, as: SessionsRegistry
   alias PtcRunnerMcp.TestSupport.MemorySoak
+  alias PtcRunnerMcp.TestSupport.SoakHelpers
 
   @moduletag :soak
   @moduletag timeout: :infinity
 
   setup do
-    stop_sessions_processes()
-    SessionsConfig.set(%{enabled: true})
-    ConcurrencyGate.reset()
-    assert :ok = Sessions.ensure_started()
-
-    on_exit(fn ->
-      stop_sessions_processes()
-      SessionsConfig.reset()
-      ConcurrencyGate.reset()
-    end)
-
+    SoakHelpers.setup_sessions()
     {:ok, iters: MemorySoak.iteration_count()}
   end
 
   test "stateless turns don't grow the session GenServer", %{iters: iters} do
-    session_id = start_session()
+    session_id = SoakHelpers.start_session()
     session_pid = session_pid!(session_id)
 
     before_session = process_memory(session_pid)
@@ -65,7 +54,7 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
     IO.puts("BEFORE (stateless turns, n=#{iters}):\n#{MemorySoak.format(before)}")
 
     MemorySoak.loop(iters, fn _phase ->
-      eval_ok!(session_id, "(+ 1 2)")
+      SoakHelpers.eval_ok!(session_id, "(+ 1 2)")
     end)
 
     after_session = process_memory(session_pid)
@@ -86,11 +75,11 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
     MemorySoak.assert_flat!(before, aft, :binary, tolerance_pct: 100)
     MemorySoak.assert_atoms_per_iter!(before, aft, iters)
 
-    close_session!(session_id)
+    SoakHelpers.close_session!(session_id)
   end
 
   test "writes that accumulate user-namespace state grow linearly", %{iters: iters} do
-    session_id = start_session()
+    session_id = SoakHelpers.start_session()
     session_pid = session_pid!(session_id)
 
     before_session = process_memory(session_pid)
@@ -103,7 +92,7 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
     # per entry, give or take).
     MemorySoak.loop(iters, fn {_phase, i} ->
       program = "(def x_#{i} (str \"value-\" #{i}))"
-      eval_ok!(session_id, program)
+      SoakHelpers.eval_ok!(session_id, program)
     end)
 
     after_session = process_memory(session_pid)
@@ -128,33 +117,7 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
     # input somewhere in the eval path.
     MemorySoak.assert_atoms_per_iter!(before, aft, iters)
 
-    close_session!(session_id)
-  end
-
-  defp start_session do
-    %{"structuredContent" => sc} =
-      Tools.call(%{"name" => "ptc_session_start", "arguments" => %{}})
-
-    sc["session_id"]
-  end
-
-  defp eval_ok!(session_id, program) do
-    response =
-      Tools.call(%{
-        "name" => "ptc_session_eval",
-        "arguments" => %{"session_id" => session_id, "program" => program}
-      })
-
-    sc = response["structuredContent"]
-    assert sc["status"] == "ok", "eval failed: #{inspect(sc)}"
-    sc
-  end
-
-  defp close_session!(session_id) do
-    Tools.call(%{
-      "name" => "ptc_session_close",
-      "arguments" => %{"session_id" => session_id}
-    })
+    SoakHelpers.close_session!(session_id)
   end
 
   defp session_pid!(session_id) do
@@ -167,25 +130,5 @@ defmodule PtcRunnerMcp.ManyTurnsSoakTest do
       {:memory, m} -> m
       nil -> 0
     end
-  end
-
-  defp stop_sessions_processes do
-    stop_if_alive(SessionsRegistry)
-    stop_if_alive(PtcRunnerMcp.Sessions.Supervisor)
-  end
-
-  defp stop_if_alive(name) do
-    case Process.whereis(name) do
-      nil -> :ok
-      pid -> stop_process(pid)
-    end
-  end
-
-  defp stop_process(pid) do
-    if Process.alive?(pid) do
-      GenServer.stop(pid, :normal, 5_000)
-    end
-  catch
-    :exit, _ -> :ok
   end
 end
