+++ mcp_server/test/soak/mcp_stdio_soak_test.exs
@@ -41,93 +41,86 @@ defmodule PtcRunnerMcp.McpStdioSoakTest do
   use ExUnit.Case, async: false
 
   alias PtcRunnerMcp.Test.ReleaseRunner
-  alias PtcRunnerMcp.TestSupport.MemorySoak
 
   @moduletag :soak
   @moduletag timeout: :infinity
 
-  setup_all do
-    if ReleaseRunner.release_built?() do
-      {:ok, skip: nil}
-    else
-      {:ok,
-       skip:
-         "release binary missing — run `MIX_ENV=prod mix release --overwrite` first " <>
-           "(expected at #{ReleaseRunner.release_bin()})"}
-    end
-  end
+  @release_skip_reason "release binary missing - run `MIX_ENV=prod mix release --overwrite` first " <>
+                         "(expected at #{ReleaseRunner.release_bin()})"
 
-  test "stdio stateless eval loop returns ok for every request", %{skip: skip} do
-    if skip do
-      IO.puts("\n[SKIP] #{skip}")
-      assert true
-    else
+  if ReleaseRunner.release_built?() do
+    alias PtcRunnerMcp.TestSupport.MemorySoak
+
+    test "stdio stateless eval loop returns ok for every request" do
       run_stdio_soak()
     end
-  end
 
-  defp run_stdio_soak do
-    iters = MemorySoak.iteration_count()
+    defp run_stdio_soak do
+      iters = MemorySoak.iteration_count()
 
-    # Build the wire script: init -> N stateless eval calls -> exit.
-    {frames, total_requests} = build_frames(iters)
+      # Build the wire script: init -> N stateless eval calls -> exit.
+      {frames, total_requests} = build_frames(iters)
 
-    started_at = System.monotonic_time(:millisecond)
+      started_at = System.monotonic_time(:millisecond)
 
-    {:ok, replies, status, stderr} =
-      ReleaseRunner.run_session(frames,
-        timeout_ms: max(iters * 50, 60_000),
-        env: [
-          # Disable any noisy telemetry sinks the soak doesn't care about.
-          {"PTC_RUNNER_MCP_LOG_LEVEL", "error"}
-        ]
-      )
+      {:ok, replies, status, stderr} =
+        ReleaseRunner.run_session(frames,
+          timeout_ms: max(iters * 50, 60_000),
+          env: [
+            # Disable any noisy telemetry sinks the soak doesn't care about.
+            {"PTC_RUNNER_MCP_LOG_LEVEL", "error"}
+          ]
+        )
 
-    elapsed_ms = System.monotonic_time(:millisecond) - started_at
+      elapsed_ms = System.monotonic_time(:millisecond) - started_at
 
-    assert status in [0, :normal], """
-    Release exited abnormally: status=#{inspect(status)}
-    stderr (last 2 KB):
-    #{String.slice(stderr, max(byte_size(stderr) - 2048, 0), 2048)}
-    """
+      assert status in [0, :normal], """
+      Release exited abnormally: status=#{inspect(status)}
+      stderr (last 2 KB):
+      #{String.slice(stderr, max(byte_size(stderr) - 2048, 0), 2048)}
+      """
 
-    # Every `ptc_lisp_execute` reply should be `status: "ok"`. We grep
-    # `structuredContent.status` on the bodies — only call-result frames
-    # have it.
-    bad =
-      replies
-      |> Enum.filter(&match?(%{"result" => %{"structuredContent" => _}}, &1))
-      |> Enum.reject(fn frame ->
-        get_in(frame, ["result", "structuredContent", "status"]) == "ok"
-      end)
+      # Every `ptc_lisp_execute` reply should be `status: "ok"`. We grep
+      # `structuredContent.status` on the bodies — only call-result frames
+      # have it.
+      bad =
+        replies
+        |> Enum.filter(&match?(%{"result" => %{"structuredContent" => _}}, &1))
+        |> Enum.reject(fn frame ->
+          get_in(frame, ["result", "structuredContent", "status"]) == "ok"
+        end)
 
-    assert bad == [],
-           "#{length(bad)} eval call(s) returned non-OK status. " <>
-             "First failure:\n#{inspect(Enum.at(bad, 0), pretty: true, limit: :infinity)}"
+      assert bad == [],
+             "#{length(bad)} eval call(s) returned non-OK status. " <>
+               "First failure:\n#{inspect(Enum.at(bad, 0), pretty: true, limit: :infinity)}"
 
-    IO.puts("""
-    Stdio soak:
-      iterations:  #{iters}
-      requests:    #{total_requests}
-      elapsed_ms:  #{elapsed_ms}
-      replies:     #{length(replies)}
-      per-iter ms: #{Float.round(elapsed_ms / max(iters, 1), 2)}
-    """)
-  end
+      IO.puts("""
+      Stdio soak:
+        iterations:  #{iters}
+        requests:    #{total_requests}
+        elapsed_ms:  #{elapsed_ms}
+        replies:     #{length(replies)}
+        per-iter ms: #{Float.round(elapsed_ms / max(iters, 1), 2)}
+      """)
+    end
 
-  # ---------------------------------------------------------------------
-  # Frame construction
-  # ---------------------------------------------------------------------
+    # ---------------------------------------------------------------------
+    # Frame construction
+    # ---------------------------------------------------------------------
 
-  defp build_frames(iters) do
-    eval_frames =
-      Enum.map(1..iters, fn i ->
-        ReleaseRunner.tools_call_request(i + 1, "ptc_lisp_execute", %{
-          "program" => "(+ 1 2 3)"
-        })
-      end)
+    defp build_frames(iters) do
+      eval_frames =
+        Enum.map(1..iters, fn i ->
+          ReleaseRunner.tools_call_request(i + 1, "ptc_lisp_execute", %{
+            "program" => "(+ 1 2 3)"
+          })
+        end)
 
-    frames = [ReleaseRunner.init_request(1), ReleaseRunner.initialized_notif() | eval_frames]
-    {frames ++ [ReleaseRunner.exit_notif()], iters + 1}
+      frames = [ReleaseRunner.init_request(1), ReleaseRunner.initialized_notif() | eval_frames]
+      {frames ++ [ReleaseRunner.exit_notif()], iters + 1}
+    end
+  else
+    @tag skip: @release_skip_reason
+    test "stdio stateless eval loop returns ok for every request"
   end
 end
